requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
pytest
aws-cdk-lib>=2.99.0
aws-cdk.aws-lambda-python-alpha>=2.99.0-alpha.0
constructs>=10.0.0
boto3
